+++ app/static/js/rpghome.js
@@ -16,12 +16,13 @@ $.getJSON('/'+home+'/sheets', function(data) {
 });
 */
 
-function createButtonGroup (id) {
+function createButtonGroup (name, id) {
     var btnG = '<div class="sheet-stat button"><a class="btn btn-secondary btn-block" href="/'+home+'/edit/'+id+'" role="button">Edit</a></div>';
-    return btnG+= '<div class="sheet-stat button-last"><button class="btn btn-danger btn-block" name="delete" onclick="deleteSheet(\''+id+'\')">Delete</button></div>';
+    return btnG+= '<div class="sheet-stat button last"><button class="btn btn-danger btn-block" name="delete" onclick="deleteSheet(\''+name+'\',\''+id+'\')">Delete</button></div>';
 }
 
-function appendRowPB (data){
+function appendRow (data){
+    $('#no-sheets').remove();
     var rowString = '<div id=row_'+data['id']+' class="sheet"><div class="block">';
     rowString += '<div class="wrap sheet-stats"><a class="sheet-show" href="/'+home+'/sheet/'+data['id']+'"><div class="wrap justify-content-left">';
     rowString += '<div class="sheet-stat name">'+formatName(data['name'])+'</div>';
@@ -30,7 +31,7 @@ function appendRowPB (data){
     rowString += '<div class="sheet-stat level">'+data['level']+'</div>';
     rowString += addSheetHitPoints(data['current_hp'],data['max_hp']);
     rowString += '</div></a></div>';
-    rowString += createButtonGroup(data['id']);
+    rowString += createButtonGroup(data['name'], data['id']);
     rowString += '</div></div>';
     $('.player-sheets').append(rowString);
 }
@@ -60,11 +61,14 @@ function createNewSheet () {
     url: '/'+home+'/create',
     dataType: 'json',
     success: function(data) {
-        appendRowPB(data['sheets']);
+        appendRow(data['sheets']);
     }});
 };
 
-function deleteSheet(id) {
+function deleteSheet(name, id) {
+    if (!window.confirm("Are you sure you want to delete: "+name+" ?" )) {
+        return false
+    };
     $.ajax({
     type: "DELETE",
     url: '/'+home+'/delete/'+id,
@@ -75,13 +79,22 @@ function deleteSheet(id) {
 };
 
 
-$.ajax({
+
+/*
+When the document is ready, query the backend to get all character sheets
+*/
+$( document ).ready(function() {
+    $.ajax({
     type: "GET",
     url: '/'+home+'/sheets',
     dataType: 'json',
     success: function(data) {
-        $.each(data['sheets'], function(index, element) {
-        appendRowPB(element);
-        });
+        if (data['sheets'].length == 0){
+            $('.player-sheets').append('<div id="no-sheets" class="justify-content-center"><p>You have no characters saved!</p></div>')
+        } else {
+            $.each(data['sheets'], function(index, element) {
+            appendRow(element);
+            });
+        };
     }
-});
+})});
